[refactor] features 폴더 생성
