+++ src/features/forecast/utils/getWeatherCondition.ts
